% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Title myclt}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{n}

\item{iter}{iter}

\item{a}{a}

\item{b}{b}
}
\value{
return sum sample
}
\description{
Title myclt
}
\examples{
n <- 5
iter <- 5
myclt(n,iter,a=0,b=5)
}
