% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximun Likelyhood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function to be applied}

\item{x}{first parameter}

\item{param}{second parameter}

\item{...}{third parameter}
}
\value{
max likelyhood
}
\description{
Maximun Likelyhood
}
\examples{
x = 10
lfun = 1
param = 50
mymaxlik=function(lfun,x,param,...)
}
