% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{normal max likelihood estimator}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{vector of data}

\item{mu}{mean of population}

\item{sig}{standard hdviation of population}

\item{...}{variables for the plot}
}
\value{
list of values for max likelihood
}
\description{
normal max likelihood estimator
}
\examples{
x=1:5
mu = 15
sig = 10
mymlnorm=function(x,mu,sig,...)

}
