% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l5function.R
\name{mysample}
\alias{mysample}
\title{Mysample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{time}{}
}
\value{
a my sample barplot
}
\description{
Mysample
}
\examples{
n=1000
time=1
iter=1
}
