% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nticket.R
\name{ntickets}
\alias{ntickets}
\title{Title: ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{the probability a plane will be truly overbooked}

\item{p}{probability of a "show"}
}
\value{
returns a list of nd, nd, N, p, and gamma as well as two plots
}
\description{
Title: ntickets
}
\examples{
N <- 100
gamma <- 0.05
p <- 0.95
ntickets(N, gamma, p)

}
