% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scatterhist}
\alias{scatterhist}
\title{A function that creates a scatterplot histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{x axis}

\item{y}{y axis}

\item{xlab}{x axis label}

\item{ylab}{y axis label}
}
\value{
returns a scatterplot histogram
}
\description{
A function that creates a scatterplot histogram
}
